ai_backend/api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
asyncpg>=0.29.0
httpx>=0.25.0
aiohttp>=3.9.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pandas>=2.1.0
matplotlib>=3.8.0
numpy>=1.24.0
bcrypt>=4.1.2
email-validator>=2.1.1
reportlab>=4.0.0
apscheduler>=3.10.0
zipfile36>=0.1.3
openai>=1.3.0
